requirements.txt
numpy==1.16.2
Pillow==7.1.2
